restructure go sdk layout
