+++ examples/go/go.mod
@@ -1,0 +1,11 @@
+module github.com/usemoss/moss/examples/go
+
+go 1.22.2
+
+require github.com/usemoss/moss/sdks/go/sdk v0.0.0
+
+require github.com/usemoss/moss/sdks/go/bindings v0.0.0 // indirect
+
+replace github.com/usemoss/moss/sdks/go/sdk => ../../sdks/go/sdk
+
+replace github.com/usemoss/moss/sdks/go/bindings => ../../sdks/go/bindings
