+++ sdks/go/sdk/client_test.go
