+++ sdks/go/sdk/read.go
